+++ backend/srp-base/src/app.js
@@ -68,6 +68,9 @@ const clothesRoutes = require('./routes/clothes.routes');
 const notesRoutes = require('./routes/notes.routes');
 // base events domain route
 const baseEventsRoutes = require('./routes/baseEvents.routes');
+// cron domain route
+const cronRoutes = require('./routes/cron.routes');
+=======
 // coordsaver domain route
 const coordsaverRoutes = require('./routes/coordsaver.routes');
 
@@ -191,6 +194,8 @@ app.use(notesRoutes);
 app.use(coordsaverRoutes);
 // mount base events routes
 app.use(baseEventsRoutes);
+// mount cron routes
+app.use(cronRoutes);
 
 // mount phone routes
 app.use(phoneRoutes);